requirements.txt
django==3.2.15
djangorestframework==3.13.1
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.1
python-decouple==3.6
mysqlclient==2.1.1
elasticsearch==7.17.6
elasticsearch-dsl==7.4.0